Update documentation and move into angular job into angular shared
